attempt to get the test discovered
